Basic clean up fixing variable names and commenting classes

+++ Assets/Scripts/GameState/BeeResources.cs
@@ -1,8 +1,8 @@
 public struct BeeResources {
   public enum Type { Honey, Beeswax, Nectar, RoyalJelly, Pollen }
-  public float honey;
-  public float beeswax;
-  public float royalJelly;
-  public float nectar;
-  public float pollen;
+  public int honey;
+  public int beeswax;
+  public int royalJelly;
+  public int nectar;
+  public int pollen;
 }
